Move models to own module, fix manifest bug.
