+++ src/llmbo/llmbo.py
@@ -2,7 +2,6 @@ import json
 import logging
 import os
 import time
-from dataclasses import dataclass
 from datetime import datetime
 from typing import Any, Dict, List, Literal, Optional, Type
 from uuid import uuid4
@@ -12,69 +11,9 @@ from botocore.exceptions import ClientError
 from dotenv import load_dotenv
 from pydantic import BaseModel, ValidationError
 
+from .models import Manifest, ModelInput, ToolChoice
+
 logger = logging.getLogger(__name__)
-
-
-@dataclass
-class Manifest:
-    totalRecordCount: int
-    processedRecordCount: int
-    successRecordCount: int
-    errorRecordCount: int
-    inputTokenCount: Optional[int]
-    outputTokenCount: Optional[int]
-
-
-@dataclass
-class ToolChoice:
-    type: Literal["any", "tool", "auto"]
-    name: Optional[str] = None
-
-
-@dataclass
-class ModelInput:
-    """Configuration class for AWS Bedrock model inputs.
-
-    This class defines the structure and parameters for model invocation requests
-    following AWS Bedrock's expected format.
-
-    See https://docs.aws.amazon.com/bedrock/latest/userguide/model-parameters-anthropic-claude-messages.html
-
-    Attributes:
-        messages (List[dict]): List of message objects with role and content
-        anthropic_version (str): Version string for Anthropic models
-        max_tokens (int): Maximum number of tokens in the response
-        system (Optional[str]): System message for the model
-        stop_sequences (Optional[List[str]]): Custom stop sequences
-        temperature (Optional[float]): Sampling temperature
-        top_p (Optional[float]): Nucleus sampling parameter
-        top_k (Optional[int]): Top-k sampling parameter
-        tools (Optional[List[dict]]): Tool definitions for structured outputs
-        tool_choice (Optional[ToolChoice]): Tool selection configuration
-    """
-
-    # These are required
-    messages: List[dict]
-    anthropic_version: str = "bedrock-2023-05-31"
-    max_tokens: int = 2000
-
-    system: Optional[str] = None
-    stop_sequences: Optional[List[str]] | None = None
-    temperature: Optional[float] = None
-    top_p: Optional[float] = None
-    top_k: Optional[int] = None
-
-    tools: Optional[List[dict]] | None = None
-    tool_choice: Optional[ToolChoice] = None
-
-    def to_dict(self):
-        result = {k: v for k, v in self.__dict__.items() if v is not None}
-        if self.tool_choice:
-            result["tool_choice"] = self.tool_choice.__dict__
-        return result
-
-    def to_json(self):
-        return json.dumps(self.to_dict())
 
 
 VALID_FINISHED_STATUSES = ["Completed", "Failed", "Stopped", "Expired"]
